requirements.txt
pandas==1.5.2
Nasdaq-Data-Link==1.0.4
pandas-datareader==0.10.0
python-dotenv==0.21.0
yfinance==0.2.12
psycopg2==2.9.5